backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
openai==1.12.0
python-jose==3.3.0
requests==2.31.0
beautifulsoup4==4.12.3 